xz: Use automatic word wrapping for help texts
This is more complex code and less flexible at formatting the
messages too. However, hopefully this is worth it to make it
easier for translators to handle the messages. (Very roughly
half of translators don't consistently verify that the help
messages are neatly aligned and stay within 80 columns when
using manual word wrapping.)