xz: Add a function to return the directory part of the filename.
Add a function which returns the directory part of the filename. This is
either the name up the last deliminator like '/' or simple '.' if it is
current directory.
The function reuses the inner parts of has_dir_sep() for the
deliminator. The inner parts are moved to get_dir_sep() in order to the
pointer while has_dir_sep() still returns the bool.

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>