xz: Sync the written file to disk before removing the original.
If the file is compressed (or decompressed) then the original file is
deleted by default after the written file is closed. All this is fine.

If the machine crashes then it is possible that the newly written file
has not yet hit the disk while the previous delete operation might have.
In that case neither the original file nor the written file is
available. To avoid this scenario, sync the file, followed the directory
(to ensure that the file is part of the directory).

This may cause the `xz' command to take a bit longer because it now
waits additionally until the data has been written to disk.

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>