FIXME *** Build: Add support for git_commit_info.sh
Using $(RECURSIVE_TARGETS) might apply to too many targets. At the same
time, it isn't enough: Each library/program or file that depends on
git_commit_version.h should ensure that the file is up to date without
trying to update it in parallel (and then corrupting it).
Now if one configures the tree and runs "make -C src", the header
isn't generated and the build fails.