Release resources even for non-debug builds.
This change makes the applications actually free the associated resources even for non-debug builds.

Not all platforms implement shared libraries in the way that the library allocator links to the parent application's heap. Not releasing resources can thus lead to actual memory leaks, which can be hundreds of megabytes in case of xz tools.