liblzma: Add a new .xz decoder flag LZMA_TELL_BLOCK_END
It makes lzma_code() return LZMA_BLOCK_END after every .xz Block.
It can be useful in special cases where the boundaries of Blocks
have a meaning to the application. For example, if using the
seekable decoder, one can seek by Block number and then decode
exactly one Block by detecting LZMA_BLOCK_END.

Even if one knows the uncompressed size of the Block, the benefit
of waiting for LZMA_BLOCK_END is that then one knows that the
integrity check has been verified. A decoder might have provided all
output from the Block, but if the last input bytes of the Block
haven't been provided to the decoder, it cannot have verified the
integrity check.