DRAFT *** Build: Autodetect unaligned access on LoongArch
*** FIXME: Untested. CMake support is missing, and I'm not sure if it
*** even is straighforward to implement.

According to [1] sections 7.4, 8.1, and 8.2, desktop and server
processors support fast unaligned access (-mno-strict-align is
the default), but embedded systems likely don't. There currently
(GCC 15.2) is no preprocessor macro on LoongArch to detect if
-mstrict-align or -mno-strict-align is in effect. This uses ugly
heuristics to detect which of the flags is in effect.

(CMake 4.1 added more CMAKE_<LANG>_COMPILER_ARCHITECTURE_ID values
to detect the arch in a cleaner manner, but 4.1 is too new to require
for now.)

[1] https://github.com/loongson/la-softdev-convention/blob/v0.2/la-softdev-convention.adoc

Thanks-to: lichenggang <lichenggang@deepin.org>
Thanks-to: Xi Ruoyao
Closes: https://github.com/tukaani-project/xz/pull/186